requirements.txt
Django==3.2
django-ninja==0.17.0
pip==21.2.4
